chore(analytics): Simplify analytics testing (#22235)
Provide an alternative implementation to the one suggested in #21680 which is
much simpler. The problem we are trying to resolve is that when QAs or even
developers need to analyze events sent to MixPanel they cannot pinpoint exactly
which events originated from their testing Status installation. Even if events
end up in the test project (which is not as noisy as prod MixPanel), it’s still
cumbersome to test if there are other people concurrently generating events
(which can be the case while we are testing releases).

The solution is to allow us to filter events based on the so called MixPanel
User ID, which in the case of Status, is a UUID generated once and which varies
per installation. This PR adds a new option in the Advanced Settings behind a
toggle, therefore the feature is only available in PR builds because feature
flags are only accessible if ENABLE_QUO_PREVIEW is 1.

Additional notes

The original solution suggested in the issue
https://github.com/status-im/status-mobile/issues/21680 is more flexible but
more involved, where clients would be able to specify their own UserID, which in
the future would be useful to support tracking certain events more anonymously
by not tying them to the global installation ID (User ID).

Steps to test

1.  Create at least one profile or log in to an existing one.
2.  Enable usage tracking in Settings > Privacy and security.
3.  Go to Settings > Feature flags and enable analytics > copy-user-id.
4.  Log out so that the app initialization kicks in.
5.  Log in and go to Settings > Advanced settings and press on the
    option Copy analytics User ID to copy the User ID to the clipboard.
6.  Verify events are being sent to MixPanel (see screenshot below as an
    example on how to filter).

- status-go PR: https://github.com/status-im/status-go/pull/6385
- Fixes https://github.com/status-im/status-mobile/issues/21680